requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==4.2.4
certifi==2024.2.2
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==15.1.0
contourpy==1.2.1
cryptography==41.0.3
cycler==0.12.1
daphne==4.0.0
Django==4.1.13
django-allauth==0.62.1
django-cors-headers==4.3.1
django-googledrive-storage==1.6.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
Flask==3.0.3
fonttools==4.51.0
google-api-core==2.10.2
google-api-python-client==2.125.0
google-auth==1.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mongoengine==0.28.2
msgpack==1.0.5
mysqlclient==2.1.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
Pillow==10.0.0
promise==2.3
protobuf==4.25.3
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==4.6.2
pyOpenSSL==23.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.3
redis==5.0.0
requests==2.31.0
rsa==4.9
seaborn==0.13.2
service-identity==23.1.0
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.2.4
text-unidecode==1.3
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
zope.interface==6.0
